+++ maintenance/views/__init__.py
@@ -1,0 +1,13 @@
+"""views package (split from the original monolithic views.py).
+
+Submodules are re-exported here so `from . import views; views.X`
+and `from <app>.views import X` keep working unchanged.
+"""
+from .helpers import *  # noqa: F401,F403
+from .activities import *  # noqa: F401,F403
+from .schedules import *  # noqa: F401,F403
+from .activity_types import *  # noqa: F401,F403
+from .imports_exports import *  # noqa: F401,F403
+from .timeline import *  # noqa: F401,F403
+from .debug import *  # noqa: F401,F403
+from .reports import *  # noqa: F401,F403
